Simplify Muzzle side-classes to make them easier to unload and remove the need to modify the Instrumenter class (#3854)
